+++ lib/jsonapi/record/destroyable.rb
@@ -11,7 +11,7 @@ module JSONAPI
         # @param record [JSONAPI::Record::Base]
         # @return [JSONAPI::Record::Base]
         def destroy(record)
-          response_document = JSONAPI::Client.delete(individual_uri(record.id), default_headers)
+          response_document = JSONAPI::SimpleClient.delete(individual_uri(record.id), default_headers)
 
           case response_document
           when Types::Failure
@@ -24,7 +24,7 @@ module JSONAPI
         end
 
         # @param record [JSONAPI::Record::Base]
-        # @raise [JSONAPI::Client::UnprocessableEntity] if destroy fails.
+        # @raise [JSONAPI::SimpleClient::UnprocessableEntity] if destroy fails.
         # @return [JSONAPI::Record::Base]
         def destroy!(record)
           raise_exception_when_errors { destroy(record) }